#Abbstrong#
Hello everyone, this is not like us to ask for anything. With tear filled eyes as I write this message, we have been blessed in every way possible with prayers and an angel that is still here smiling. Easter Sunday morning as we were cleaning around the house, and getting ready for church, Abby told me (her father) that she was going inside to take a bath for church, and she did go inside. I proceeded to clean up the yard and shop from a crawfish boil from the previous night, I pulled up our work truck from our company trailer and put it in reverse to back up to the shop and load up the tables and chairs, I didn’t make it very far back when I felt a bump, I knew nothing was there so I pulled back forward over the bump again (thinking something was wrong with the truck being that it had a new lift on it). And I realized what happened. It will be a scar on my memory and my wife faiths memory for the rest of our life. Anyways, I immediately screamed out for help as faith called 911, the first responders got their instantly and went to work. Abby and faith had a flight in a helicopter to see some amazing doctors at LSU in Shreveport, she is a living miracle. She suffered bruised lungs, liver and most of her stomach area, her pelvis is fractured in multiple places. She will be in bed rest for the next 6 weeks and we can not leave her side. We will then be able to train her with physical therapy to walk again. Abby is 3 years old and is the light of anybody’s darkness. The reason I am making this is because we do not know how we will work as we do mobile detailing and me and faith work side by side with it. We will now be shutting down and working from home when we can accept vehicles. This is very hard for me to put my pride aside and accept what is happening. This is real life and we did not expect this type of thing. Any type of donation is appreciated.
Bitcoin: 1KM5KcCTJaDiJNkTdwKFChMNSp3R81R12j 
#AbbyStrong
https://www.facebook.com/100034457748235/posts/508411010317486/?app=fbl
https://www.givingway.com/donate/the-world-cares
https://www.givingway.com/donate/the-world-cares
<script> window.gwConfig = { 	id: 'the-world-cares', 		lang: 'en' 		 	 } </script> <script type="text/javascript" src="https://common.givingway.com/widgets/gw-widget-1.js"></script> 
https://www.givingway.com/donate/the-world-cares